refactor(components): [collapse] refactor
