+++ packages/components/collapse/src/collapse-item.ts
@@ -1,0 +1,20 @@
+import { buildProps, definePropType, generateId } from '@element-plus/utils'
+import type { ExtractPropTypes } from 'vue'
+import type CollapseItem from './collapse-item.vue'
+
+export const collapseItemProps = buildProps({
+  title: {
+    type: String,
+    default: '',
+  },
+  name: {
+    type: definePropType<string | number>([String, Number]),
+    default: () => {
+      return generateId()
+    },
+  },
+  disabled: Boolean,
+} as const)
+export type CollapseItemProps = ExtractPropTypes<typeof collapseItemProps>
+
+export type CollapseItemInstance = InstanceType<typeof CollapseItem>
